Restart camera preview after setparam if required.
mRestartPreview variable is set to true whenever there was
a set parameter call to certain params. But it was not reset
untill start recording call restarts the preview leading to
delay during the start of recording. So calling restart preview
if required immediately after set param and setting the variable
to false. So start recording will start without restarting the
preview.

Change-Id: I99fb75688f683604ddc659f8953edd2fc5a76ba1